requirements.txt
Click==7.0
click-completion==0.5.2
colorama==0.4.3
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.18.1
pandas==0.25.3
psycopg2-binary==2.8.4
python-dateutil==2.8.1
python-dotenv==0.10.4
pytz==2019.3
shellingham==1.3.1
six==1.14.0
SQLAlchemy==1.3.12
typer==0.0.8
